split src files
