+++ src/provider.js
@@ -1,0 +1,52 @@
+import { createContext, createElement, useMemo } from 'react';
+
+import { batchedUpdates } from './batchedUpdates';
+
+// context
+const warningObject = {
+  get dispatch() {
+    throw new Error('Please use <ReduxProvider store={store}>');
+  },
+  get getState() {
+    throw new Error('Please use <ReduxProvider store={store}>');
+  },
+};
+export const ReduxStoreContext = createContext(warningObject);
+
+// patch store with batchedUpdates
+const patchReduxStore = (origStore) => {
+  if (!batchedUpdates) return origStore;
+  const listeners = [];
+  let unsubscribe;
+  const subscribe = (listener) => {
+    listeners.push(listener);
+    if (listeners.length === 1) {
+      unsubscribe = origStore.subscribe(() => {
+        batchedUpdates(() => {
+          listeners.forEach(l => l());
+        });
+      });
+    }
+    return () => {
+      const index = listeners.indexOf(listener);
+      listeners.splice(index, 1);
+      if (listeners.length === 0) {
+        unsubscribe();
+      }
+    };
+  };
+  return {
+    ...origStore,
+    subscribe,
+  };
+};
+
+// provider
+export const ReduxProvider = ({ store, children }) => {
+  const patchedStore = useMemo(() => patchReduxStore(store), [store]);
+  return createElement(
+    ReduxStoreContext.Provider,
+    { value: patchedStore },
+    children,
+  );
+};
