+++ src/use-redux-dispatch.js
@@ -1,0 +1,8 @@
+import { useContext } from 'react';
+
+import { ReduxStoreContext } from './provider';
+
+export const useReduxDispatch = () => {
+  const store = useContext(ReduxStoreContext);
+  return store.dispatch;
+};
